telegram-bot/requirements.txt
python-telegram-bot
firebase-admin
requests